backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
httpx==0.26.0
python-dotenv==1.0.0
sentence-transformers==2.3.1
pdfplumber==0.10.3
numpy==1.26.3

# LLM providers (install as needed)
openai==1.10.0
google-generativeai==0.3.2
mistralai==0.1.2
